finalizado entity, service y controller de turno

+++ clinicaOdontologica/src/main/java/com/backend/clinicaOdontologica/service/ITurnoService.java
@@ -9,8 +9,8 @@ public interface ITurnoService {
     TurnoDto guardarTurno(Turno turno);
     List<TurnoDto> listarTodos();
 
-    TurnoDto buscarTurnoPorId(int id);
+    TurnoDto buscarTurnoPorId(Long id);
     TurnoDto actualizarTurno(Turno turno);
-    void eliminarTurno(int id);
+    void eliminarTurno(Long id);
 
 }
